+++ Cryptocurrencies.ROI.Calculator/Domain/CryptoCurrencies/Ethereum.cs
@@ -28,7 +28,7 @@ namespace Cryptocurrencies.ROI.Calculator.Domain
 
         }
 
-        public decimal CalculateROIinDays(ComputingDevice device,  int powerConsumptionPerHour, decimal energyPricePerKWH, decimal currencyRate)
+        public decimal CalculateROIinDays(ComputingDevice device,  int powerConsumptionPerHour, decimal energyPricePerKWH, double currencyRate)
         {
             
             decimal coinsPerDay = (((device.HashPower * 1000000 * BlockRewards) / NetworkDifficulty) * 3600 * 24);
