+++ Cryptocurrencies.ROI.Calculator/Domain/CryptoCurrencies/ICalculateROI.cs
@@ -9,7 +9,7 @@ namespace Cryptocurrencies.ROI.Calculator.Domain
     public interface ICalculateROI
     {
 
-        public decimal CalculateROIinDays(ComputingDevice device, int powerConsumptionPerHour, decimal energyPricePerKWH, decimal currencyRate);
+        public decimal CalculateROIinDays(ComputingDevice device, int powerConsumptionPerHour, decimal energyPricePerKWH, double currencyRate);
     
     }
 }
